[EAGLE-754] Refactor - move ignored test class to corresponding module
