+++ eagle-core/eagle-query/eagle-query-base/src/test/java/org/apache/eagle/query/TestListQueryCompiler.java
@@ -17,7 +17,7 @@
 /**
  * 
  */
-package org.apache.eagle.service.generic;
+package org.apache.eagle.query;
 
 import org.apache.eagle.log.entity.meta.EntityDefinition;
 import org.apache.eagle.log.entity.meta.EntityDefinitionManager;
